A generic INSDC scanner with EMBL and GenBank subclasses, for use with the scanner/consumer model used in Bio.GenBank
This code is NOT intended for direct use, it is to be used by Bio.GenBank to parse
GenBank files and by Bio.SeqIO to parse GenBank and EMBL files.

Apart from a minor change in the handling of the new line white space in features
qualifier values, the switch should be transparent to existing Bio.GenBank users.