TreeIO: fixed parse() closing handles too early
Py2.4 doesn't allow yield inside try/finally; using return instead causes the
file handle to close immediately, but parse() needs to keep the handle open as
long as trees remain in the file, so we use yield to make it work. This means an
exception could leave an open file handle in Jython if parse() is called with a
filename instead of a handle.