Exposed the '_al_start' and '_al_stop' private variables of the SeqRecord objects, created a new public 'annotations' dictionary in the Alignment object
The '_al_start' and '_al_stop' variables that were private to the SeqRecord class are now copied into the public
dictionary 'annotations' of the SeqRecord class with the keys 'start' and 'end'.

The Alignment object has a private '_annotations' dictionary. The code now creates a public 'annotations' dictionary and
copies the value corresponding to the 'sw_expect' key from '_annotations' to 'annotations' in order to make it
publicly accessible.