Phylo: fixed the bug in root_at_midpoint
Most trees don't have a branch length assigned to the root, so depths()
was ignoring this value without a noticeable problem. But when
root_at_midpoint() calls depths() on child clades, the leading branch
length should be included, at it wasn't.

Solution: include the root branch length in depths(). In
root_at_midpoint(), adjust the calculation of 'remainder' for the root
branch length since the root branch will not be traversed.