Inter-block coordinate pairs of fragmented HSPs are now a property
This commit adds two new properties to the HSP object:
`query_inter_coords` and `hit_inter_coords`. They are coordinates of the
regions between each HSPFragment, in query and hit sequences each.

Previously, the Exonerate parsers computes these properties and sets it
as an instance attribute. Since this addition makes the computation
uneeded, it has been deleted.

Also, the attributes `ner_ranges` and `intron_ranges` of the Exonerate
parsers are now merged into one, in the new inter-block ranges.