Cleanup unused import statements
As in Nick's pull request excluding Bio.GA and Bio.Graphics changes.

Peter

--

I'm trying to clean up a few parts of the codebase. Big things first:
Pylint complains about these unused imports.

I hate commits that touch lots of different parts of the codebase,
but this is all the same logical issue.

It'd be great if someone else could test this, especially the Graphics
components that depend on reportlab.

Signed-off-by: Nick Semenkovich <semenko@alum.mit.edu>