autopep8 on rest of SeqIO (with minor tweaks)
I noticed autopep8 didn't like a hash comment line just before a
function definition, and would insert two blank lines between them.
In most cases this was a redundant 'This is a generator function!'
comment which is probably not needed now that Biopython makes
quite a lot of usage of generator functions.