Update Bio/PDB/DSSP.py
There is a bug in DSSP module, when some lines whitout text appears in DSSP output (header). I'm solved this problem with a easy modification in line 121:

            if len(sl)>1 and sl[1] == "RESIDUE":

This len(sl)>1 solves the problem ;)

The next is the output of my dssp (DSSP 2.0.4). The line numper 3 of 5gat-1.dssp_2 has length 1:

['====', 'Secondary', 'Structure', 'Definition', 'by', 'the', 'program', 'DSSP,', 'CMBI', 'version', 'by', 'M.L.', 'Hekkelman/2010-10-21', '====', 'DATE=2012-11-13', '.']
['REFERENCE', 'W.', 'KABSCH', 'AND', 'C.SANDER,', 'BIOPOLYMERS', '22', '(1983)', '2577-2637', '.']
['.']
                                                                                                                               .
A lot of thanks!!! Excuse me for the English

Thanks :)