Switch BLAST XML indexing to be line based.
This now *requires* that the <Iteration> marker be the first tag on a
line (which is how the NCBI BLAST XML files look), although in theory
an edited XML file could remove new line whitespace.

The reason for this change is to facilitate indexing BGZF compressed
BLAST XML files.

In principle we can still use the original approach, but determining
the start offset for a BGZF compressed BLAST XML file is far harder.