Remove raw text length calculation from hmmer3-text
Hmmer3-text parsing & indexing requires the header to be present. This means the
raw query result text needs to be appended to the header. If the query result
text length is calculated, appending is not possible since SearchIO.index_db
directly returns the text in the given range, without any modification.

Note that raw text modification may be possible, but we need to meddle with
the base `index_db` class to do this ~ which may be too messy.