Fix doctest helper function
The idea of the helper function is to allow you to edit any module,
for example Bio/Align/__init__.py, and then run its doctests in situ.
This should work from the Biopython root, Bio, Bio/Align, Bio/Tests/,
etc.

The old code was wrongly defaulting to the __file__ value of the
Bio/_utilts.py but that could give you the location of an installed
Biopython. This now defaults to the current directory.