Issue a warning when translating partial codons
See mailing list dicussion. Some felt this was useful functionality
and perhaps worthy of another optional argument to the translate
functions/methods, but they are already quite complicated. Using
a warning for now, and an exception later is to help migrate old
code which may be using this (e.g. ORF finding code).

This will requires explicitly padding/trimming the sequence to be
a multiple of three before translation.