SeqIO.GenBank: survive empty feature qualifiers
When a GenBank file contains an empty qualifier for a key that already
contained a qualifier, the `feature_qualifier()` function fails with

AttributeError: 'NoneType' object has no attribute 'replace'

because the code doesn't return from the if value is None check designed
to catch this case.
This patch fixes the AttributeError by just returning.

Signed-off-by: Kai Blin <kai.blin@biotech.uni-tuebingen.de>