GenBank: Give better error output on invalid qualifiers
When encountering a multiline qualifier that doesn't contain an = sign
and doesn't correctly close the quotation marks, the error message
"TypeError: unsupported operand type(s) for +: 'NoneType' and 'str'"
is less than helpful. Catch the None type and give a better error
message by raising a StopIteration.

Also add a unittest-based test case for this, as I couldn't figure out
how to properly test this with the old-style tests.

Signed-off-by: Kai Blin <kai.blin@biotech.uni-tuebingen.de>