Use built in next() function in test_NCBIXML.py
This change is enough to run this test file unmodified
under Python 3 where iterators don't have a next method
(see the new __next__ method instead). The next function
is built in in Python 2.6 onwards

We probably need to do this elsewhere in Biopython, this
particular test file is an interesting one to start with
because running 2to3 on it takes a long time.