Disable installation under Python 3.1 (unittest import broke)
Disabling Python 3.1 is due to the recent change to relative
imports (via the 2to3 import fixer) and our unit tests breaking.
Something about the unittest and/or __import__ function breaks,
and since Python 3.1 is so old I'm not going to try to solve it.

Although we're officially supporting Python 3.3 onwards, this
will still work under Python 3.2 which is useful in the short
term for testing PyPy3 (which so far only targets Python 3.2).