SeqIO._lazy: integration to SeqIO.parse, SeqIO.index_db. Added key_function.
Multiple additions to make SeqIO._lazy work with both parse and index_db.
Key function support was added to allow users to modify record ids with
a key function as currently supported by index_db. More tests were added
to work with this. A duplicate fasta file (of f002) was added to support
a test that required failure when identical records were accessed in
index_db.