NCBIXML.py: added coverage calculation
* Bio/Blast/NCBIXML.py (_get_master_coverage): Add calulation logic for the percentage query coverage per subject (new property `alignment.qcovs').
(_get_hsp_coverage): Add calulation logic for the percantage query coverage per HSP (new property `hsp.qcovhsp').
(length, intersection, combination, merge_range_list): Helper functions for implementing percantage coverage logic.

This is a new feature for the NCBIXML parser where the resulting Blastresult object now contains additional information on the query coverage. This information is not contained in the Blast XML format but rather is calculated while parsing. It makes sence to include this information as it can be present in the Blast tabular output (generated by the blast programms or with blast_formatter) and allows for easier filtering of the resultsets.