KGML_vis.py: Changed hexdarken to darken, and added show_reaction_entries option
The hexdarken function expected a hex-encoded colour, but was only used
for foreground colouring, so assignments of ReportLab colours slipped
past the tests. I renamed the function to darken() [there should be no
downstream side-effects] and changed all calls in this file,
appropriately. I also modified darken() so that it tests for hex- or RGB
colour encoding, and returns a darkened colour of the same type.

The show_reaction_entries option was necessary because it was not
possible to entirely remove ortholog elements from KEGG maps
without also removing reaction_entries elements. The previous
code did not allow for reaction_entries to be removed, only their
labels.