MOTIF: implemented instances.search with difflib.SequenceMatcher; however this will be discarded in a future commit
I implemented instances.search using difflib.SequenceMatcher, but the problem is that difflib only returns one match (http://stackoverflow.com/questions/28716510/how-to-get-multiple-matches-with-difflib-sequencematcher)

In any case this approach returns all the possible matches of the motif on the sequence, which for large sequence may be troublesome. Need to find a better solution!