Allow ABI parser to pick up extended data
This commit disables the check that the key be "known". At the expense
of users possibly encountering keys that don't have a built-in label,
the format is future-proofed against manufacturers adding additional
keys/value pairs to the files, and users not being able to access that
data.

The test case diff showcases that we were obscuring data from the users
that they might want to access.