Fix bug in refmat(p, q)
Instead of normalizing a copy of the p and q vectors, refmat would
normalize p and q themselves!
As a conseauence, calling rotmat on two vectors would normalize them,
which is a huge side-effect that I am pretty sure is a bug and not a
feature.

NB: This bug was discovered while fixing the previous one, as I was
curious to know while calling rotmat([0,0,0],[0,1,0]) would
**set** the first vector to [Nan, Nan, Nan]