Fix hmmer3-text bug where table hit description has preference over full descriptions
Previously, the hit description in the hit table always overwrites the full
description. This doesn't matter when they are the same, but the description
in the tables has a maximum length. If the full description exceeds the length,
a truncated version is shown in the table.

This change ensures that the full description is always used.