Update docstring for Bio.NaiveBayes: scale=False
The original code in Bio.NaiveBayes (current line 96) shows the statement:

```python
if scale:
    do_something()
```

The funcion does scaling of probabilities if ``scale`` is not False (by
default no scaling is done).
This implies that the parameter ``scale`` is intended to be checked for True or
False values. However, the function definition in line 59 shows that ``0`` is
being used as indication of False.

I thought that intention should be clearer by using booleans (False, True)
instead of using ``0`` and ``any other value that will test for True``.

The modification to Bio.NaiveBayes was done after writing tests covering the
original workflow of the module.

I updated the docstring for Bio.NaiveBayes.calculate to point out this change.