replaced "if variable is True" for "if variable"
PEP8 recommends against using the following statements:

if variable == True:

if variable is True:

PEP8 recommends using instead "if variable:".

In this case, the variable "shift" in Bio.codonalign is meant to indicate True
or False. Thus, I replaced the line 167 to "shift = False" as None is not quite
the oposite of True.