'dumb_consensus' conservation fix at AlignInfo.py
Hello,
Currently, 'dumb_consensus' counts atoms just after skipping gaps, so if there are 100 sequences in an alignment, and in a given position there is an insertion in only one sequence (meaning 1 residue and 99 gaps), the position will be treated as 100% conserved instead of just 1%, and will pass any threshold. Since I'm looking for strings with different degrees of conservation, this is creating strings with some residues that are not as conserved as expected.

To fix it I moved the atom counting just after the length check, before the gap skipping, and did the same for 'gap_consensus' for consistency. Hope you can merge it.

Best regards,