Updated BioSQL sqlite3 schema to include foreign keys
Fix for GitHub issue #741 by changing the sqlite3 database schema
to use foreign keys. A new test has been added in that explicitly
tests to make sure associated data with a bioentry is removed.

To accommodate the use of foreign keys the connection to an sqlite
database must set the PRAGMA, which is now done automatically when
opening the database.

The upstream BioSQL schema for sqlite3 has also been updated, see
https://github.com/biosql/biosql/pull/3