Updated sqlite3 schema to include foreign keys
Fix for biopython/biopython#741 by changing the database schema
for sqlite3 to use foreign keys. A new test has been added in that
explicitly tests to make sure associated data with a bioentry is
removed.

To accommodate the use of foreign keys the connection to an sqlite
database must set the PRAGMA, which is now done automatically when
opening the database.

The sqlite schema for testing has been updated to be identical to
the version in biosql/biosql#3 that is a pull request for the main
biosql repository