Support for GAF 2.1 in the UniPort.GOA GAF parser.
Fixed a bug in GOA GAF parser. The parser defaulted to GAF 1.0
if it didn't recognizes the GAF version. Since there are now 3
GAF versions (2.1, 2.0 and historically 1.0), the code makes an
explicit choice when parsing the gene associate file header and
raises an exception if the GAF version is not recognized.

Squashed commit from GitHub pull request #818.