remove assertion
Seems to be giving me lots of trouble with checking that the taxon
id is either an int or a long. I don't think this assertion is
even necessary since it has to be an integer in the database so I
can't imagine what error that assertion is really checking for.