Some KEGG REST queries now return String not Byte in Python3
See issue #832.

Python 3 makes a distinction between string types (String and
Byte) that was not made in Python 2. This gave unintuitive results
with print() and SeqIO.read() in Python3. This short fix returns
String rather than Byte handles (as with Bio.Entrez), for all
requests through the KEGG REST API, other than those for image
data, which is still returned as a Byte handle.