Correct CAI Implementation and Multiple Genetic Codes
Fixes an incorrect implementation of the CAI formula as defined by Sharp and Li. They state "that if a certain codon is never used in the reference set then the CAI for any other gene in which that codon appears becomes zero. To overcome this problem we assign a value of 0.5 to any [codon] that would otherwise be zero." Also adds support for multiple genetic codes.