Outline CONTRIBUTING file for GitHub Pull Requests.
GitHub should show this to people submitting pull requests,
and ought to take advantage of markdown (CONTRIBUTING.md)
or reStructuredText (CONTRIBUTING.rst) in preference to
plain text (CONTRIBUTING).

I have followed our current pattern with README (included
in the source releases) and symlink README.rst for GitHub
(not included in the source releases).