Handle malformed GenBank locations like order(462..470,480..485,)
The trailing comma should not be there, and hints at
a data-loss somewhere in the production of these files.
e.g. record NC_024258.1 in the current version of
ftp://ftp.ncbi.nih.gov/refseq/release/plastid/plastid.1.genomic.gbff.gz

See Biopython GitHub issue #937.

With this change a warning is issued, and the trailing
comma removed.