Downgrade single-entry GenBank compound locations to a warning
e.g. order(20991..20999,) in the record for NC_024258.1
in the current version of this file:

ftp://ftp.ncbi.nih.gov/refseq/release/plastid/plastid.1.genomic.gbff.gz

This location (and others in the same record) has a
trailing comma (suggestive of dataloss upstream, see
previous commit which will remove this and give a
warning), and only one part.

See Biopython GitHub issue #937.

With this change a warning is issues, and the problematic
feature gets a location of None.