Adressed technical issues mentioned in review, and added an additional version of one test.
I changed the ValueErrors in the tests to self.asserts.
One could even comment out the first 2 asserts in `test_full_id_is_updated_chain` and
`test_full_id_is_updated_residue`.
The first assert mostly serves to document my assumptions about the data loaded.
The second verifies that there is indeed caching taking place, so we need to invalidate the full_id.
The third and fourth assertion test the desired behaviour of the public API. They are the main assertions of the tests.