Stop the MMTF parser treating NUL bytes as altlocs
MMTF uses NUL bytes to indicate that an atom does not have an
alternative location. The MMTF parser was passing these NUL bytes
("\x00") directly into the "StructureBuilder.init_atom" method, which
expects a space (" ") to indicate that an atom has no altloc.

This was not picked up by unit tests because the mmtf files bundled with
biopython incorrectly specified a space as the altloc. This commit
replaces those files with files downloaded directly from the RCSB.