Include FH spacer line in EMBL output
This is described in the (current) EMBL flat file format specification,
quoting ftp://ftp.ebi.ac.uk/pub/databases/embl/doc/usrman.txt it says:

3.4.15  The FH Line
The FH (Feature Header) lines are present only to improve readability of
an entry when it is printed or displayed on a terminal screen. The lines
contain no data and may be ignored by computer programs. The format of these
lines is always the same:
     FH   Key             Location/Qualifiers
     FH
The first line provides column headings for the feature table, and the second
line serves as a spacer. If an entry contains no feature table
(i.e. no FT lines - see below), the FH lines will not appear.