Mark residues as disordered when parsing MMCIF
Squashed commit of pull request #1014.

This commit updates MMCIFParser to match the behaviour of PDBParser (and
the documentation) when parsing point mutations. This is done by calling
the "init_residue" of StructureBuilder when the residue name changes, as
well as when the residue ID changes.

None of the existing MMCIF files contain point mutations, so I have
added 3JQH.cif for parsing.