Non-descriptive failure in python 3 Affy.CelFile.read_v4
A piece of code like this:

import Bio.Affy.CelFile as CELFile
with open("CELFileV4.CEL"), "r") as f:
    record = CELFile.read_v4(f)

fails in a non-descriptive way with UnicodeDecodeError in python 3. Instead, an error should be raised, which informs the user what went wrong, in this case, opening the file in a wrong mode.

This has already been in place with the version-detecting function read.

This pull request proposes a unified approach, which shares error handling between read and read_v4.

In python 2 there is no problem, the record is read in correctly.