Unix sort the CONTRIB file.
We already sort names in the NEWS file alphabetically by
first name, while the CONTRIB file was traditionally done
by surname. This was inconsistent, and also is harder to
get right.

This change sorts the CONTRIB file using Unix sort, which
means by first name, and enforces this via the TravisCI
Tox setup. Ideally it would be added to any official git
pre-commit hook as well.