Added support for deuterium as an element.
For some proteins in the PDB, it is important for researchers to
distinguish between hydrogen isotopes.  Those researchers will often
populate the element field with 'D' for deuterium.  This usage is
in accordance with IUPAC provisional recommendation IR-3.3.2, but
there is no reference to deuterium in the atom_weights table in the
IUPACData.py file so these values are not recognized.

The value used for the atomic weight of deuterium was taken from
http://www.ciaaw.org/hydrogen.htm.

For the purposes of testing this change, one of the HETATM lines in the
a_structure.pdb file was modified to reference a deuterium atom instead
of the carbon atom it originally represented.  This required minor
changes to the ParseTest class.  A specific test for this change was
added to the Atom_Element class.