Preserve order in SeqIO and SearchIO index, index_db
This is simpler to think about, and also means looping over the
index to iterate over a file (or files) will be done in the
disk-order, which will be faster.

(Previously iteration over the index would happen in the sorted
order of the index key hashes, essentially random, thus there
would be a lot of disk random access jumping back and forth
through the file to parse the records in arbitrary order.)

Internally the SeqIO/SearchIO index_db function will now use
explicit sorting via the file number and offset so that the
keys and iteration order reflect that on disk.

Internally the SeqIO/SearchIO index function will now use an
OrderedDict on Python 3.5 or older.

Python 3.6 has an ordered dict, but this was declared an
implementation detail only, rather than a language standard.

PyPy has always had an ordered dict.

Python 3.7 onwards has an ordered dict as part of the
language definition, see:

https://mail.python.org/pipermail/python-dev/2017-December/151283.html

We do not in general want to use OrderedDict which is designed
for different use cases (making lots of changes to the dictionary
content), see:

https://mail.python.org/pipermail/python-dev/2017-December/151266.html