use a single RNG rather than a new one for each use.
Random is random, but there's a bug under Python 3.6 for Windows where multiple RNGs created in a short window (about 15ms) all have the same initial seed, and hence be correlated. While it's not likely to be a serious problem, there's no need to have multiple RNGs in the first place. Instead, simply share the same generator for all three tasks.