MAFINDEX_VERSION = 2 (inclusive end coordinates).
The use of inclusive coodinates for the sqlite MAX index
is an attempt to fix issues discussed in the following pull requests:
https://github.com/biopython/biopython/pull/504
https://github.com/biopython/biopython/pull/1086

A test has been added in `test_MafIO_index.py` to check that the number
of
MAF alignment blocks returned when querying for a single position will
be 1 at the boundary between blocks (it could be 2 with the previous MAF
index, which is a bug).

The indices built with end-exclusive coordinates will not be compatible
with the present version.