Explicitly set long_description to our README.rst
Based on the advice here:

https://docs.python.org/2/distutils/packageindex.html#pypi-package-display
https://docs.python.org/3/distutils/packageindex.html#pypi-package-display

This will hopefull solve the failure of PyPI to show our
README file in recent releases,

https://pypi.python.org/pypi/biopython/1.66 - worked
https://pypi.python.org/pypi/biopython/1.67 - worked
https://pypi.python.org/pypi/biopython/1.68 - no readme shown
https://pypi.python.org/pypi/biopython/1.69 - no readme shown

I don't think anything we changed had an impact here.

Both 1.67 and 1.68 had RST formatted text in README, and there
was a symlink README.rst as well.

As of Biopython 1.69 we have an RST formatted file README.rst,
and a symlink README (no extension) as well.