Tests: Handle recent PhyML versions without ".txt" output suffix (closes #919)
The PhyML version packaged in Ubuntu 17.04 (phyml 3:3.2.0+dfsg-7ubuntu1)
drops the suffix ".txt" from output file names. See: stephaneguindon/phyml#39

Also, the release date is not longer included in the version string, so
checking for the string "20" (current century) does not detect a successful
execution of "phyml --version". As a fallback, check for the string
"PhyML" in the output.