TreeConstruction: Add 'skip_letters' parameter to DistanceCalculator (#491)
When scoring with a distance matrix other than 'identity', the gap and
stop characters '-' and '*' are skipped & not counted toward the total
score. With the 'identity' model, the default for DistanceCalculator,
these characters are retained and compared. That's potentially
surprising, so this patch exposes the 'skip_letters' keyword argument in
the constructor so that the scoring/skipping behavior can be changed for
any model.

In a future release, it would be better to set the default value of
skip_letters=('-', '*') and have 'identity' use this default value like
the other models do. For now, we'll leave a shim for backward
compatibility.