Fix use of backslash-t in BWA example (D301).
Passing a literal tab character in a command line string
is problematic (and would require quoting which our wrapper
does not attempt to do). Instead, BWA itself understands the
backslash-t notation to insert a tab in the read group string.

This means when constructing the argument in Python, we need
to escape the backslash - or use a raw string.

This also fixes the doctest output which was showing white
space in the expected output of the command line as a string.

Bonus: This is the last D301 issue from pydocstyle/flake8