SeqRecord: Don't have a hard dependency on BioSQL (fixes #1471)
Currently, SeqRecord expects BioSQL to always be importable. That is not true on Debian packages,
where the -sql package is needed to have BioSQL around.

When BioSQL is not available, the instance in question should not be a DBSeqRecord, though.
That means the else branch of that check should be safe.

Signed-off-by: Kai Blin <kblin@biosustain.dtu.dk>