Bio.File.as_handle: support pathlib.Path objects and more (PEP 519)
The logic changes for this are quite small: instead of checking whether
we think `handleish` should be passed to `open`, just try that first. If
that succeeds, yield the new file handle; otherwise yield the object
as-is.

With this change in behavior, it is no longer BioPython's responsibility
to know whether a certain object can be passed to `open`. This allows a
user to use a `pathlib.Path` object (or `os.DirEntry` returned from
`os.scandir`) under Python >= 3.6. Behavior in previous versions of
Python is unchanged.