Tests.test_File.AsHandleTestCase: change assertion in test_stringio
The docstring of Bio.File.as_handle (currently) contains:

    All other inputs are returned, and are *not* closed.

As such, we should really test object identity here instead of object
equality; I don't think a StringIO object will ever compare unequal to
itself, but the intention of this test is to verify that we get the same
object back from `File.as_handle` if it's not something that can be
understood by `open`.

I considered replacing this test with a `test_arbitrary_object` test
case that passed an `object()` to `File.as_handle`, but that seems like
unnecessary code churn.