Update operators for Residue
The == operator for residues currently seems to only compare identity
of the `id` field which changes behaviour to v1.70. 
The same applies for the other operators as residues with identical "id" fields are considered
equal. This e.g. causes NeighborSearch to report the same residue pair twice if the id field is identical.
Last the  __le__ and __lt__ were identical which I assume is not intended.