Remove residue specific comparison and perform entity comparison on full ID
comparing the full ID yields the same result as comparing first the parents
and then the entities themselves and for residue fields in the ID are evaluated
in the same order as the spelled out comparisons implemented before.
Thus a customized test seems superfluous. While this eliminates the changes in
Neighbor search reported, comparisons might not yield the expected result but should be
consistent with behaviour of version 1.70 in python 2