Fix bug in mmCIF parsing for mid-line comments
Comments in mmCIF files are only started if the "#" character is at the
beginning of a line or preceded by whitespace:
https://www.iucr.org/resources/cif/spec/version1.1/cifsyntax#lex
This means that an unquoted valid token can contain a "#".

This commit adds a check to the Bio.PDB.MMCIF2Dict.MMCIF2Dict._splitline
method to account for this. Previously, a "#" character would always
start a mid-line comment, which broke the parsing of (e.g.) 1EMB at the
line "_entity_src_gen.pdbx_host_org_vector TU#58". I have also added
some tests to hopefully prevent regressions here.

I introduced this bug in 781441b :)