Only use OrderedDict in to_dict on old Python
Python 3.6 has an ordered dict, but this was declared
an implementation detail only rather than a language
standard.

PyPy has always had an ordered dict.

Python 3.7 onwards has an ordered dict as part of the
language definition, see:

https://mail.python.org/pipermail/python-dev/2017-December/151283.html

We do not in general want to use OrderedDict which is
designed for different use cases (making lots of changes
to the dictionary content), see:

https://mail.python.org/pipermail/python-dev/2017-December/151266.html