Add cif-seqres and cif-atom SeqIO drivers
This commit adds the 'cif-seqres' and 'cif-atom'
drivers to the Bio.SeqIO module. These are implemented by the
'CifSeqresIterator' and 'CifAtomIterator' functions in the
'Bio.SeqIO.CifIO' module.

Most of the logic for the 'cif-atom' driver is identical to the
'pdb-atom' driver, as they both read from a Bio.PDB Structure object.
The shared code has been refactored into the 'AtomIterator' function in
the 'Bio.SeqIO.structure_io' module.

The cif-seqres iterator reads three categories: '_pdbx_poly_seq_scheme',
which contains the actual sequence; '_struct_ref', which contains
database cross-references; and '_struct_ref_seq', which is linked to
_struct_ref via the ref_id field and contains the PDB and chain ID of
the cross-reference.

Since so much of the test code is shared between the pdb-* and cif-*
SeqIO readers, the 'test_SeqIO_PdbIO.py' file has been renamed to
'test_SeqIO_StructureIO.py', and now contains the tests for both the
pdb-* and cif-* readers. For the sake of consistency, I have also added
2XHE.cif and a_structure.cif to the Tests/PDB directory.

Shared test code is implemented in a superclass returned from a
function in order to convince unittest to run the child test suites but
not the parent suite.