GenBank/EMBL: Downgrade bad origin wrapping to a warning
See issue #1608 - Since Biopython 1.71 and pull request #1497,
GenBank or EMBL files for circular genomes (eg length 7000)
have triggered an error for origin wrapping locations expressed
as the shorthand like '6800..100' rather than the expected but
verbose 'join(6801..700,1..100)' style.

This change downgrades the ValueError exception to a warning,
with the bad location left as None.

In future, we could automatically upgrade such locations into
the expected form (with a different warning).

Alternatively, some think we should remove the check in the
SeqFeature location objects which verifies start <= end, which
is essential for all the functionality like length, iteration,
etc.