Avoid creating new class instance in argument defaults.
This avoids a flake8-bugbear warning,

B008 Do not perform calls in argument defaults. The call is performed only once at function definition time. All calls to your function will reuse the result of that definition-time call. If this is intended, assign the function call to a module-level variable and use that variable as a default value.

We already have a module-level variable with a shared
instance of the generic protein alphabet - so use that.