Set PDBList local_pdb default at initialisation time
Previously default would be the current directory at the
time of the import:

$ python -c "import os; print(os.getcwd()); from Bio.PDB.PDBList import PDBList; os.chdir('/tmp'); print(PDBList().local_pdb); os.chdir('/Users/peter/'); print(PDBList().local_pdb)"
/Users/peter/repositories/biopython
/Users/peter/repositories/biopython
/Users/peter/repositories/biopython

Now:

$ python -c "import os; print(os.getcwd()); from Bio.PDB.PDBList import PDBList; os.chdir('/tmp'); print(PDBList().local_pdb); os.chdir('/Users/peter/'); print(PDBList().local_pdb)"
/Users/peter/repositories/biopython
/private/tmp
/Users/peter

Spotted via flake8-bugbear,

Bio/PDB/PDBList.py:86:58: B008 Do not perform calls in argument defaults. The call is performed only once at function definition time. All calls to your function will reuse the result of that definition-time call. If this is intended, assign the function call to a module-level variable and use that variable as a default value.