Standard license header for parse_pdb_header.py
Also moved some of the header comments into the docstring, and
belatedly added Kristian Rother to the CONTRIB.rst listing.

Note the comment:

"license: same as Biopython, read LICENSE.TXT from current Biopython release."

This can likely be used to switch this file to dual-licensing
with the 3-clause BSD license if-and-when we manage to dual-license
the rest of Biopython, even if we don't have all the individual
contributor's explicit agreement. See issue #898.