Ingnore flake8-bugbear B007 in Scripts/
This was failing locally, but not on TravisCI, which is odd:

$ flake8 Scripts/
Scripts/query_pubmed.py:45:14: B007 Loop control variable 'arg' not used within the loop body. If this is intended, start the name with an underscore.
Scripts/xbbtools/nextorf.py:198:26: B007 Loop control variable 'length' not used within the loop body. If this is intended, start the name with an underscore.
Scripts/Performance/biosql_performance_read.py:23:5: B007 Loop control variable 'junk_id' not used within the loop body. If this is intended, start the name with an underscore.
Scripts/Performance/biosql_performance_read.py:39:5: B007 Loop control variable 'junk_id' not used within the loop body. If this is intended, start the name with an underscore.
Scripts/debug/debug_blast_parser.py:131:9: B007 Loop control variable 'err_file' not used within the loop body. If this is intended, start the name with an underscore.
Scripts/debug/debug_blast_parser.py:131:19: B007 Loop control variable 'err_line' not used within the loop body. If this is intended, start the name with an underscore.
Scripts/debug/debug_blast_parser.py:131:43: B007 Loop control variable 'err_text' not used within the loop body. If this is intended, start the name with an underscore.
Scripts/debug/debug_blast_parser.py:240:14: B007 Loop control variable 'arg' not used within the loop body. If this is intended, start the name with an underscore.

We currently ignore B007 in Bio/ etc.