Update AbiIO parser to preserve bytes for element codes 2, 18, and 19
Although the original specification mentions that the string values of
element 2, 18, 19 should be representable in the ascii range, some of
these tags may contain bytes value that represent a wider range (e.g.
non-ascii utf-8). This is because these tags are used to store free-form
user annotations (e.g. the CMNT-1 tag), which may well include non-ascii
characters.

In the interest of making the parser preserve more robust (i.e. not
breaking when encountering these non-ascii values), this commit updates
the parser so that it keeps the original byte values when parsing.

This does not include all tags however. There are still tags whose value
will be parsed into strings, since SeqRecord requires them to be
strings. These are tags storing the sequence ID and description, for
example. In this commit, _bytes_to_string (meaning decoding using utf-8)
is still called to transform those tags' values.