Sample problematic GCG MSF file from IPD-IMGT/HLA database
https://github.com/ANHIG/IMGTHLA/blob/3320/msf/DOA_prot.msf
as of commit 16c09d89398603dcf653cc5476f857f1a21c1d9d
(14 August 2018), i.e. from version 3.32.0 of the DB.

This file has a discrepancy in the alignment length in the
main header (62) versus most of the records (250).

If we accept there should be 250 columns, then the entry
for the last sequence is only 62 without gap padding
(and indeed, has blank lines in the last three blocks).

The initial Biopython GCG MSF parser will accept this file
but with two warnings.