Debug PDBIO.py
There was an issue in the PDBIO.py (before even this set of corrections)
that caused the PDBIO module to output incorrect .pdb files. The problem
was that the charge field in _get_atom_line() was overwritten in the
case of .pdb files with charge = None which caused the problem. This
issue is now fixed.

In addition, fixed another minor bug in PDBIO and a linting error in
Atom.py