Four codons per row; blank gaps for stop codons
Historically the table dictionary was given with four
codons per row, but the stop codons were skipped over.

This uses four potential codons per row with gaps as
needed for stop codons, i.e. always 16 rows.

This makes it visually easier to check the tables,
although the order is not the same as that on the
NCBI website's table view.